Refactored some classes and methods
