+++ src/main/java/qupath/ext/gdcnn/tasks/AnnotationExportTask.java
@@ -12,7 +12,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javafx.concurrent.Task;
-import qupath.ext.gdcnn.listeners.ProgressListener;
+import qupath.ext.gdcnn.entities.ProgressListener;
 import qupath.ext.gdcnn.utils.Utils;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.QuPathGUI;
