+++ src/main/java/qupath/ext/gdcnn/tasks/TilerTask.java
@@ -9,7 +9,7 @@ import org.slf4j.LoggerFactory;
 
 import javafx.collections.ObservableList;
 import javafx.concurrent.Task;
-import qupath.ext.gdcnn.listeners.ProgressListener;
+import qupath.ext.gdcnn.entities.ProgressListener;
 import qupath.ext.gdcnn.utils.Utils;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.QuPathGUI;
